feat(auth): add rbac in app

+++ src/features/posts/server/actions/delete-posts.ts
@@ -2,8 +2,9 @@
 
 import { db } from "@/db";
 import { post } from "@/db/schema";
+import { hasPermissions } from "@/features/auth/server/actions/has-permissions";
 import { auth } from "@/lib/auth";
-import { inArray } from "drizzle-orm";
+import { and, eq, inArray } from "drizzle-orm";
 import { revalidateTag } from "next/cache";
 import { headers } from "next/headers";
 
@@ -18,12 +19,34 @@ export async function deletePosts(postIds: string[]) {
     };
 
   try {
-    const res = await db.delete(post).where(inArray(post.id, postIds));
-    revalidateTag("posts");
-    return {
-      success: true,
-      message: `Deleted ${res.rowCount} post(s)`,
-    };
+    if (await hasPermissions(userSession.user.id, { post: ["delete:any"] })) {
+      const res = await db.delete(post).where(inArray(post.id, postIds));
+      revalidateTag("posts");
+      return {
+        success: true,
+        message: `Deleted ${res.rowCount} post(s)`,
+      };
+    } else if (
+      await hasPermissions(userSession.user.id, { post: ["delete:own"] })
+    ) {
+      const res = await db
+        .delete(post)
+        .where(
+          and(
+            inArray(post.id, postIds),
+            eq(post.authorId, userSession.user.id),
+          ),
+        );
+      revalidateTag("posts");
+      return {
+        success: true,
+        message: `Deleted ${res.rowCount} post(s)`,
+      };
+    } else
+      return {
+        success: false,
+        message: "You do not have the permissions to perform this action",
+      };
   } catch {
     return {
       success: false,
